Refactor Operation + Fragment Builder Tests
